# docs/swagger.yaml
basePath: /
definitions:
  client.Error:
    properties:
      code:
        type: integer
      error:
        type: string
    type: object
...
  client.Success:
    properties:
      ok:
        type: boolean
    type: object
...
  model.FullTicketInfo:
    properties:
      buyTime:
        type: string
      company:
        type: string
      endPoint:
        type: string
      endTime:
        type: string
      id:
        type: integer
      startPoint:
        type: string
      startTime:
        type: string
      userId:
        type: integer
    type: object
  model.ReportFlights:
    properties:
      notFulFilled:
        items:
          $ref: '#/definitions/model.FullTicketInfo'
        type: array
      previously:
        items:
          $ref: '#/definitions/model.FullTicketInfo'
        type: array
      previouslyAndNotFulFilled:
        items:
          $ref: '#/definitions/model.FullTicketInfo'
        type: array
    type: object
...
  user.TimeDiapason:
    properties:
      endTime:
        type: string
      startTime:
        type: string
    type: object
  user.User:
    properties:
      name:
        example: Name
        type: string
      patronymic:
        example: Patronymic
        type: string
      surname:
        example: Surname
        type: string
    type: object
host: localhost:80
info:
  contact: {}
  description: Test task for SmartWay
  title: Swagger of API
  version: "1.0"
paths:
...
  /user/delete/{userId}:
    delete:
      description: delete a user
      parameters:
      - description: user id
        in: path
        name: userId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/client.Success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/client.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/client.Error'
      tags:
      - user
  /user/get/{userId}:
    post:
      consumes:
      - application/json
      description: Get a report of user for diapason
      parameters:
      - description: user id
        in: path
        name: userId
        required: true
        type: integer
      - description: query params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/user.TimeDiapason'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ReportFlights'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/client.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/client.Error'
      tags:
      - user
  /user/new:
    post:
      consumes:
      - application/json
      description: create a new document
      parameters:
      - description: query params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/user.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/client.Success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/client.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/client.Error'
      tags:
      - user
  /user/update/{userId}:
    patch:
      consumes:
      - application/json
      description: update user document
      parameters:
      - description: user id
        in: path
        name: userId
        required: true
        type: integer
      - description: query params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/user.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/client.Success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/client.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/client.Error'
      tags:
      - user
swagger: "2.0"